HOTFIX/build: 빌드용 파일 정리
